ari-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ari-core"
version = "0.5.0"
description = "Artificial Research Intelligence - Core Framework"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "litellm>=1.0",
    "mem0ai>=0.1",
    "pydantic>=2.0",
    "typer>=0.12",
    "rich>=13.0",
    "pyyaml>=6.0",
    "paramiko>=3.0",
]

[project.scripts]
ari = "ari.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.hatch.build.targets.wheel]
packages = ["ari"]
